Update info to also have fixed commands
